refactor: clean up sosotalk page structure
