+++ src/features/sosotalk-write/model/use-sosotalk-write-page.ts
@@ -1,0 +1,108 @@
+'use client';
+
+import { useState } from 'react';
+
+import { useRouter } from 'next/navigation';
+import { toast } from 'sonner';
+
+import {
+  uploadSosoTalkPostImage,
+  useCreateSosoTalkPost,
+  useGetSosoTalkPostDetail,
+  useUpdateSosoTalkPost,
+} from '@/entities/post';
+
+import type { SosoTalkPostSubmitPayload } from './sosotalk-post-editor.types';
+
+interface UseSosoTalkWritePageParams {
+  editPostId?: number;
+}
+
+export function useSosoTalkWritePage({ editPostId }: UseSosoTalkWritePageParams) {
+  const router = useRouter();
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const isEditMode = editPostId != null;
+  const { data, isLoading, isError } = useGetSosoTalkPostDetail(editPostId);
+  const createPostMutation = useCreateSosoTalkPost();
+  const updatePostMutation = useUpdateSosoTalkPost();
+
+  const resolveCreateImageUrl = async (
+    payload: SosoTalkPostSubmitPayload
+  ): Promise<string | undefined> => {
+    if (payload.imageFile) {
+      return uploadSosoTalkPostImage(payload.imageFile);
+    }
+
+    return payload.displayImageUrl || undefined;
+  };
+
+  const resolveUpdateImageValue = async (payload: SosoTalkPostSubmitPayload): Promise<string> => {
+    if (payload.imageFile) {
+      return uploadSosoTalkPostImage(payload.imageFile);
+    }
+
+    return payload.displayImageUrl;
+  };
+
+  const handleCreateSubmit = async (payload: SosoTalkPostSubmitPayload) => {
+    if (isSubmitting || createPostMutation.isPending) {
+      return;
+    }
+
+    setIsSubmitting(true);
+
+    try {
+      const image = await resolveCreateImageUrl(payload);
+      const createdPost = await createPostMutation.mutateAsync({
+        payload: {
+          title: payload.title,
+          content: payload.contentHtml,
+          image,
+        },
+      });
+
+      router.push(`/sosotalk/${createdPost.id}`);
+    } catch {
+      toast.error('게시글 등록에 실패했어요. 다시 시도해 주세요.');
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
+
+  const handleEditSubmit = async (payload: SosoTalkPostSubmitPayload) => {
+    if (editPostId == null || editPostId <= 0 || isSubmitting || updatePostMutation.isPending) {
+      return;
+    }
+
+    setIsSubmitting(true);
+
+    try {
+      const image = await resolveUpdateImageValue(payload);
+
+      await updatePostMutation.mutateAsync({
+        postId: editPostId,
+        payload: {
+          title: payload.title,
+          content: payload.contentHtml,
+          image,
+        },
+      });
+
+      router.push(`/sosotalk/${editPostId}`);
+    } catch {
+      toast.error('게시글 수정에 실패했어요. 다시 시도해 주세요.');
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
+
+  return {
+    data,
+    isEditMode,
+    isError,
+    isLoading,
+    isSubmitting,
+    handleCreateSubmit,
+    handleEditSubmit,
+  };
+}
