+++ src/widgets/sosotalk/ui/sosotalk-main-page/model/index.ts
@@ -1,0 +1,2 @@
+export { createSosoTalkMainPageQueryParams } from './sosotalk-main-page.utils';
+export { useSosoTalkMainPage } from './use-sosotalk-main-page';
